pyproject.toml
[project]
name = "basic-reflection-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.15",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.9",
    "python-dotenv>=1.2.1",
]
